sandbox/pyproject.toml
[project]
name = "django_haystack_opensearch_demo"
version = "0.1.0"
authors = [
    {name = "Caltech IMSS ADS", email = "imss-ads-staff@caltech.edu"},
]
maintainers = [
    {name = "Chris Malek", email = "cmalek@caltech.edu"},
]
requires-python = ">= 3.11"
description = "Demo app for django_haystack_opensearch"
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Framework :: Django :: 5.2",
    "Typing :: Typed",
]
dependencies = [
    "django_haystack_opensearch",
    "django-braces>=1.15.0",
    "django-crispy-forms>=2.3",
    "crispy-bootstrap5>=2024.10",
    "django-environ>=0.11.2",
    "django-theme-academy>=0.3.11",
    "django-wildewidgets>=1.2.3",
    "django-xff>=1.3.0",
    "django-compressor>=4.5.1",
    "django-crequest>=2018.5.11",
    "django-debug-toolbar>=4.4.2",
    "gunicorn >= 22.0.0",
    "ipython>=7.27.0",
    "mysqlclient==2.2.6",
    "redis~=5.2.1",
    "structlog>=24.2.0",
    "tabulate==0.9.0",
    "django-sass-processor>=1.4.1",
    "libsass>=0.23.0",
    "rich>=13.9.4",
    "django-extensions>=4.1",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "pytest-env>=0.8.1",
    "pytest-timeout>=2.1.0",
    "pytest-xdist>=3.5.0",
    "pytest-randomly>=3.12.0",
    "pytest-django>=4.11.1",
]

[tool.uv.sources]
django_haystack_opensearch = {path = "./django_haystack_opensearch.tar.gz"}

[tool.uv]
python-preference = "only-system"

[tool.setuptools]
# ...
# By default, include-package-data is true in pyproject.toml, so you do
# NOT have to specify this line.
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]